+++ dev/src/dev.clj
@@ -40,3 +40,7 @@
 
 (defn q [sql]
   (jdbc/query (db) sql))
+
+(comment
+ (q "select (1+(random() * (max(id)-1)))::int from drills")
+ :rcf)